add common subpackage without libdparse
adds ddoc.unhighlight and ddoc.types modules for more useful usage.
Unhighlight can be used to emit embedded D code as-is, instead of
parsing it and (potentially) causing issues as well as needing a
dependency on libdparse.

The ddoc.types module now contains the Comment struct that was in
ddoc.comments and has a new parseUnexpanded call, which puts the split
sections directly into the comment, without expanding, which is commonly
used for example in D-Scanner to just check if a comment is `ditto` or
contains some section with a given name.